+++ efb/src/route/mod.rs
@@ -19,18 +19,20 @@ use std::rc::Rc;
 use log::{debug, trace, warn};
 
 use crate::error::Error;
-use crate::fp::Performance;
+use crate::fp::{ClimbDescentPerformance, LegPerformance};
 use crate::measurements::Speed;
 use crate::nd::*;
-use crate::{VerticalDistance, Wind};
+use crate::VerticalDistance;
 
 mod accumulator;
 mod leg;
+mod leg_fuel;
 mod profile;
 mod token;
 
 pub use accumulator::TotalsToLeg;
 pub use leg::Leg;
+pub use leg_fuel::LegFuel;
 pub use profile::{AirspaceIntersection, VerticalPoint, VerticalProfile};
 use token::Tokens;
 pub use token::{Token, TokenKind};
@@ -88,42 +90,31 @@ impl Route {
     /// idents read from the navigation data `nd`.
     pub fn decode(&mut self, route: &str, nd: &NavigationData) -> Result<(), Error> {
         debug!("route decode: {:?}", route);
+        self.clear();
         self.tokens = Tokens::new(route, nd);
-        self.legs.clear();
 
-        // clear values relevant during parsing of all tokens
-        self.origin.take();
-        self.destination.take();
-        self.takeoff_rwy.take();
-        self.landing_rwy.take();
-
-        let mut level: Option<VerticalDistance> = None;
-        let mut tas: Option<Speed> = None;
-        let mut wind: Option<Wind> = None;
+        // the builder keeps track of level changes etc
+        let mut builder = Leg::builder();
         let mut from: Option<NavAid> = None;
         let mut to: Option<NavAid> = None;
 
         for token in &self.tokens {
             match token.kind() {
                 TokenKind::Speed(value) => {
-                    tas = Some(*value);
-                    // first speed is cruise speed
-                    if self.speed.is_none() {
-                        self.speed = Some(*value);
-                        debug!("cruise speed set to {:?}", value);
-                    }
+                    builder.tas(*value);
                 }
 
                 TokenKind::Level(value) => {
-                    level = Some(*value);
-                    // first level is cruise level
-                    if self.level.is_none() {
-                        self.level = Some(*value);
-                        debug!("cruise level set to {:?}", value);
-                    }
+                    builder.cruise(*value);
                 }
 
-                TokenKind::Wind(value) => wind = Some(*value),
+                TokenKind::LevelAtFix(value) => {
+                    builder.level_at_fix(*value);
+                }
+
+                TokenKind::Wind(value) => {
+                    builder.wind(*value);
+                }
 
                 TokenKind::Airport { arpt, rwy } => {
                     // Track for leg building
@@ -178,8 +169,7 @@ impl Route {
             match (&from, &to) {
                 (Some(from), Some(to)) => {
                     trace!("creating leg: {} -> {}", from.ident(), to.ident());
-                    self.legs
-                        .push(Leg::new(from.clone(), to.clone(), level, tas, wind));
+                    self.legs.push(builder.build(from.clone(), to.clone()));
                 }
                 _ => continue,
             }
@@ -201,6 +191,10 @@ impl Route {
     pub fn clear(&mut self) {
         self.tokens.clear();
         self.legs.clear();
+        self.origin.take();
+        self.takeoff_rwy.take();
+        self.destination.take();
+        self.landing_rwy.take();
         self.alternate.take();
     }
 
@@ -213,14 +207,17 @@ impl Route {
     ///
     /// The cruise speed or level is remove from the route by setting it to
     /// `None`.
+    #[deprecated]
     pub fn set_cruise(&mut self, _speed: Option<Speed>, _level: Option<VerticalDistance>) {
         todo!("Add/remove speed and level from the elements")
     }
 
+    #[deprecated]
     pub fn speed(&self) -> Option<Speed> {
         self.speed
     }
 
+    #[deprecated]
     pub fn level(&self) -> Option<VerticalDistance> {
         self.level
     }
@@ -232,16 +229,14 @@ impl Route {
         self.alternate = alternate;
     }
 
-    /// Returns the final leg but going to the alternate.
+    /// Diverts the last leg to the alternate.
+    ///
+    /// Returns `None` if no alternate is set or if the route is empty.
     pub fn alternate(&self) -> Option<Leg> {
-        let final_leg = self.legs.last()?.clone();
-        Some(Leg::new(
-            final_leg.from().clone(),
-            self.alternate.clone()?,
-            final_leg.level().copied(),
-            final_leg.tas().copied(),
-            final_leg.wind().copied(),
-        ))
+        let alternate = self.alternate.clone()?;
+        self.legs
+            .last()
+            .map(|final_leg| final_leg.divert(alternate))
     }
 
     /// Returns the origin airport if one is defined in the route.
@@ -276,8 +271,8 @@ impl Route {
     ///
     /// ```rust
     /// # use efb::route::Route;
-    /// # use efb::prelude::Performance;
-    /// # fn accumulate_legs(route: Route, perf: Performance) {
+    /// # use efb::prelude::LegPerformance;
+    /// # fn accumulate_legs(route: Route, perf: LegPerformance) {
     /// // Iterate through route showing progressive totals
     /// for (i, totals) in route.accumulate_legs(Some(&perf)).enumerate() {
     ///     println!("Leg {}: Total distance: {}, Total fuel: {:?}",
@@ -295,23 +290,21 @@ impl Route {
     /// [totals]: `TotalsToLeg`
     pub fn accumulate_legs<'a>(
         &'a self,
-        perf: Option<&'a Performance>,
+        perf: Option<&'a LegPerformance<'a>>,
     ) -> impl Iterator<Item = TotalsToLeg> + 'a {
         self.legs
             .iter()
             .scan(None, move |totals_to_leg: &mut Option<TotalsToLeg>, leg| {
-                // accumulate totals from previous legs
                 *totals_to_leg = Some(match totals_to_leg.as_ref() {
                     None => TotalsToLeg::new(leg, perf),
                     Some(prev) => prev.accumulate(leg, perf),
                 });
-                // the totals up to this leg
                 *totals_to_leg
             })
     }
 
     /// Returns the totals of the entire route.
-    pub fn totals(&self, perf: Option<&Performance>) -> Option<TotalsToLeg> {
+    pub fn totals(&self, perf: Option<&LegPerformance>) -> Option<TotalsToLeg> {
         self.accumulate_legs(perf).last()
     }
 
@@ -327,7 +320,7 @@ impl Route {
     /// # use efb::route::Route;
     /// # use efb::nd::NavigationData;
     /// # fn show_profile(route: &Route, nd: &NavigationData) {
-    /// let profile = route.vertical_profile(nd);
+    /// let profile = route.vertical_profile(nd, None, None);
     ///
     /// for intersection in profile.intersections() {
     ///     println!("{}: {:.1} NM to {:.1} NM",
@@ -337,8 +330,13 @@ impl Route {
     /// }
     /// # }
     /// ```
-    pub fn vertical_profile(&self, nd: &NavigationData) -> VerticalProfile {
-        VerticalProfile::new(self, nd)
+    pub fn vertical_profile(
+        &self,
+        nd: &NavigationData,
+        climb: Option<&ClimbDescentPerformance>,
+        descent: Option<&ClimbDescentPerformance>,
+    ) -> VerticalProfile {
+        VerticalProfile::new(self, nd, climb, descent)
     }
 }